Refactored jwt token generation and move to utils
